* Changed library to call by reference instead of using pointers. Makes code much cleaner. * RelayNodes renamed to RepeaterNode to distiguish this from RelayActuator. * Ack is now possible to request for all message types. Added ack-flag for some of the exposed internal messages and presentation message.

+++ libraries/MySensors/examples/LightLuxSensor/LightLuxSensor.ino
@@ -29,7 +29,7 @@ unsigned long SLEEP_TIME = 30; // Sleep time between reads (in seconds)
 BH1750 lightSensor;
 MySensor gw;
 MyMessage msg(CHILD_ID_LIGHT, V_LIGHT_LEVEL);
-int lastlux;
+uint16_t lastlux;
 
 void setup()  
 {
